Add pre-cropping of image before auto-trimming is applied.
Also deal with the situation in which the video-subsystem
displays an image because image-magick actually does not load
from a URL (so it stumped me why crop-trimming an image from a
URL does not work).

Also: don't consider warnings in image loading a problem. There
is a iCCC warning that was handled as non-image load situation, but
are actually fine to ignore (was hidden before, because it ended
up at the video subsystem).