If we use the lower block, we can't leave the bottom line in reset.
We reset the terminal at each end of line, so at the last line, the
pixel beyond the last pixel (the overhang pixel in odd height images),
will look like the standard color the terminal would show text.
We need to make sure to leave this background.
(TODO: in the last line, always switch to the upper block?)