Special handling for bottom line to stay at background color.
The bottom line in an uneven line image situation should just
be neutral. However if we by default use the lower half block to
display, its 'neutral' color will be the text terminal color, but
we'd like to have the terminal backround color.

The previous fix just made sure to set this always to black, but
that only really works if the terminal itself has a black background
color.

So the only way is to always use the upper block glyph for the
last line in the odd-row situation: set the pixels with the upper
half in character color, leave the background at the bottom in its
natural form.

This now should always work with any terminal background
(tested on light and dark backgrounds)..