Make choice of pixel character configurable; exclusively use that.
Turns out, switching to the upper block for the last line creates
enough artifacts in some terminals to be annoying. So now exclusively
use the chosen character (which requires to either start the first
or end the last row with an empty pixel line in case of odd height).

Now, the choice of character is also configurable via an
environment variable (TIMG_USE_UPPER_BLOCK).

This actually simplifies code slightly.