Use long options where we previously used letter-options with opt arg.
An optional argument on a single-letter option is a gnu extension. This
does not work as intended on BSD based systems. So make -T and -s
long options (--trim and --scroll), but also backward-compatibility
support the old options for a while.

While at it, add some other long options, such as --version and
--help.