Special case JPEG loading as we can take advantage of downsampling on load.
Create new JPEGSource and implement it with turbojpeg, which provides
a nice façade to downsample on load. The downsampled scales can only
be done in some fixed fractions, so create the smallest image that is
larger than what we'd need to scale, then scale to size. Given that this
is already downsampled, it is faster.

Since this is not using GraphicsMagick, use sws_scale() for scaling
and libexif to read exif information.

Neat side-effect: sws scale produces a more pleasent output than the
GraphicsMagick default.