Allow for fast interruption while also keeping terminal in good state.
To allow us to clean up the terminal after we get an interrupt (Ctrl-C,
SIGTERM), our signal handler is setting a flag for the rest of the
code to shut down as early as possible (e.g. an infinite animation) and
leaving the terminal in a clean state.
However, that does not work for image loading as we're using external
libraries that don't know about that flag and will not be interrupted.

This change limits the time the signal handler is active to the period
where we actually do write to the terminal and where we'd like to have a
controlled exit. Re-enable the default signal handler in-between and make
sure that in that time, we have the terminal in a good state (e.g. cursor
visible), so that any interruption coming will exit fast _and_ cleanly.

This in particular helps to quickly exit timg when there is a huge image that
is being loaded and holding up display; previously Ctrl-C resulted in a wait
until that picture was loaded, now we can exit quickly.