Reduce load on terminal emulator by only emitting changed pixels.
In typical animations, not all pixels change with each frame. We
can use that to skip over pixels that have the same color as in the
previous frame and only emit changed pixels. This can reduce the load
on the terminal as less bytes need to be parsed.

Keep a backing buffer that keeps track of the previous color of a particular
pixel-pair; count number of pixels in a line that are the same and then
horizontal move cursor to the first position we have a changed pixel-pair.