Improve background rendering best effort in case of -b none
Formulate the double row append not in terms of top line and bottom
line but rather foreground and background. This pushes the choice
to the call site of AppendDoubleRow(), but allows to work with constant
values there, as it is clear which branch is used for foreground
and background.

This allows to to add what only works in the background
branch: reset the background color to get the terminal default to
get transparency.

Rarely used as the user most certainly will use -bauto, but also
does not cost too much performance to not be implemented.