Accomodate for slow terminal query: run in background.
Only if we ever deal with transparency, we might have to wait for
the terminal background color query to return, but otherwise we
already can start decoding images while that is happening.
Usually the query is very fast (<10ms), but there also have been
cases of more than a second on a slow computer over a flaky network;
accomodate for that.