Add quarter glyph rendering (not wired up yet to main())
Add GlyphPick() implemeentation that chooses one of the 8 blocks
possible to fill four quadrants with foreground/background color
(there are 16 combinations, but only 8 equivalent taking inverted
colors into account).

Since we only have two different colors to use for four pixels, the
output is naturally only an approximation. For each of the
8 possible blocks, the colors lumped together in foreground and
background are averaged in linear color space, then the color
distance of each of the quadrants to their respective averaged
color is added up. The variant with the smallest distance is picked.

Using template parameters to make sure not more branches are
introduced at runtime for the existing half-block.

Next step: actually wire these up to main(); we need some sort
of flag there to choose (right now, it just behaves as the
original half-block rendering).