Query for Kitty Graphics capability.
Unfortunately, this requires working around a bug in Konsole: it
does not absorb the for it unknown application specific query but
spills it on the terminal.
But since Konsole also does not report its window size in pixels,
we can use that as a workaround.
(TODO: Might need to include environment variable)