Optionally use compession in pixel-accurate Kitty and iTerm2 canvas.
Both, the Kitty and iTerm2 allow to send a PNG image instead of
raw bytes. This can be a lot smaller to transfer and base64 en-/decode
at the expense of more CPU usage of timg.

Trade-off pays off when running timg remotely and bandwidth is
precious. OTOH, overall CPU-use (timg + terminal) might be worse when
playing a video locally.

Not yet documented in README or man-page. In experimental mode now.