Simplify: Don't do P6 encoding anymore, just always PNG compressed or not.
Now that we have an option to send uncompressed PNG in a choice of
bandwidth-saving 24 bits, the difference to P6 is not so much anymore.
It slower than plain P6 PNM, (due to double CRC32 inside PNG encoding), but
simplicity is worthwhile. Also some terminals (such as the hterm), can't
deal with P6.

Uncompressed sending is still a couple of times faster than compressed
sending. And both of them are not a problem on an average machine.