Reduce chance of a switched-off cursor on Ctrl-C of long image load.
We switch on the cursor between images, but with the buffered writer,
that would only make it once the next image is sent. If user cancels
while an image is loaded, that won't make it to the terminal anymore.

Send it immediately right now, but this should probably be added
to the end of Send().