Use kitty protocol by default on wezterm.
The kitty protocol seems to be working better with wezterm
currently.

The current released version 20230408 does not
work at all with iterm image protocol (
`"Error decoding image: inconsistent <..>x<..>"`), while the
current version from head does work again, but is strongly flickering
in animations ( https://github.com/wez/wezterm/issues/3882 )

The Kitty protocol is now implemented in WezTerm and it works with
animations in both, the last release version and head.

While at it, make the --compress option take an optional parameter
for the compression level (default is 0, without parameter, it is
1 and with parameter allows for 0..9).