Simplify BufferedWriteSequencer mem alloc; accept std::future
The complicated memory buffer resource management is not really
needed in real life; memory allocation and deallocation of large
chunks is well handled by the system.

Change interface to send a single argument with a tuple of
buffer and size (OutBuffer). With that, prepare the code to accept
std::future's of data. That allows us to de-couple PNG compression
in a thread in the next step.