Add sixel support using libsixel.
Choose with -psixel. This is useful for terminals such as xterm
(start with -ti vt340 option), mlterm that do support this format.
Not fully tested yet with all terminals. There is a known glitch on
Konsole, but there -piterm or -pkitty is available anyway.

This is a compile-time option (default on).