Provide Kitty capability for tmux using Unicode Placeholders
In general, this is a way to work around tmux that can't deal with
graphics protocols directly.
https://sw.kovidgoyal.net/kitty/graphics-protocol/#unicode-placeholders

The needed workaround is not pretty, but that's all we got :/

This is a preparation to output properly tmux-escaped Unicode
placeholders (not wired up yet).

The default behavior without tmux of Kitty stays as-is (no
Unicode Placeholders) to be compatible with terminals that just
implement the image sub-set of the Kitty protocol.

Issues: #95