Use Primary Device Attributes feature query to detect Sixel.
Use this as a fallback query if the primary query based on environment
variables and terminal name is not conclusive. The DA1 query might
tell us if the terminals support the Sixel protocol.

Best effort: we will still not know if the newline workaround is needed.

While at it, some smallish clean-ups, e.g. low-level memmem() is only
used in one location.

Issue #145