sixel: Implement DEC-accurate calculation of cursor position for jumping
Right now only known to be implemented by development version of
Windows Terminal.

`TIMG_SIXEL_NEWLINE_WORKAROUND` now is a bitmap that allows for the
various newline behaviors to be set

  * 1 = regular newline/carriage return choice
  * 2 = full cell jump (this new feature)
  * 3 = the OR'ed value out of 1 and 2

Until we can detect that this is the terminal we're running, requires
to set environment variable; for windows terminal, that is
`TIMG_SIXEL_NEWLINE_WORKAROUND=3`.

Issues: #145